# backend/dist/api.yml
...
openapi: 3.0.0
info:
  title: api
  description: API for project
  version: 1.0.0
...
paths:
...
  /users/favorites/addGame:
    post:
      tags:
        - users
      summary: add game to user favorites games list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                gameID:
                  type: integer
                  description: game id which the user saved in favorites
                  example: 5
      responses:
        '201':
          description: The game successfully saved as favorite
        '400':
          description: Missing one or more parameters
        '401':
          description: There is no logged in user
        '404':
          description: Game isn't exist
        '409':
          description: Game already marked as favorite
...
  /users/favorites/getGames:
    get:
      tags:
        - users
      summary: get the list of favorites games
      responses:
        '200':
          description: Results of games favorites
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FavoriteGames'
        '401':
          description: There is no logged in user
...
components:
  schemas:
...
    GamePreview:
      properties:
        gameID:
          type: integer
          example: 4
          description: the id of the game
        gameDateTime:
          type: string
          description: date and hour of the game in string format
          example: '2021-08-20T18:30:00.000Z'
        homeTeamID:
          type: integer
          description: id of home team
          example: 85
        awayTeamID:
          type: integer
          description: id of away team
          example: 939
        stadium:
          type: string
          example: BGU
          description: stadium name
...
    FavoriteGames:
      type: array
      items:
        $ref: '#/components/schemas/GamePreview'
      description: list of game previews
...